services/inventoryService/pyproject.toml
[tool.poetry]
name = "inventoryservice"
version = "0.1.0"
description = "managing and providing information about the inventory of products"
authors = ["Enrique Torres"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
odmantic = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
